+++ src/solverpy/benchmark/launcher.py
@@ -2,12 +2,13 @@ import os
 import random
 import logging
 
-from ..benchmark.path import bids
+from .path import bids
 from ..tools import human, log
 from ..task.solvertask import SolverTask
 from ..task.bar import SolvingBar, RunningBar
 from ..task import launcher 
 from .reports import markdown
+from .db.providers.solved import Solved
 
 logger = logging.getLogger(__name__)
 
@@ -32,11 +33,17 @@ def init(setup=None):
 def jobname(solver, bid, sid):
    return f"{solver}:{sid} @ {bid}"
 
-def run(solver, bid, sid, desc=None, taskdone=None, db=None, cores=4, shuffle=True, force=False, **others):
+def run(solver, bid, sid, desc=None, taskdone=None, db=None, cores=4, shuffle=True, force=False, solvedby=None, it=None, **others):
+   others = dict(others, force=force, it=it, solvedby=solvedby)
    desc = desc if desc else jobname(solver, bid, sid)
    logger.debug(f"evaluating {desc}: {jobname(solver, bid, sid)}")
    # prepare the tasks to be evaluated
-   ps = bids.problems(bid)
+   solvable = None
+   if solvedby and (it == 0):
+      solvable = Solved(bid, solvedby, solver.limits.limit).cache
+      if solvable:
+         logger.debug(f"evaluation: restricted to {len(solvable)} problems solvable by {solvedby}")
+   ps = solvable if solvable else bids.problems(bid) 
    tasks = [SolverTask(solver,bid,sid,p) for p in ps]
    logger.debug(f"evaluation: {len(tasks)} tasks scheduled")
    # check for the (cached) results in the database